Fix css blocking in playwright fixtures
